requirements.txt
chromadb>=0.5.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiohttp-retry==2.9.1
aiosignal==1.4.0
alpaca-py==0.43.2
annotated-types==0.7.0
anthropic==0.93.0
anyio==4.13.0
attrs==26.1.0
beautifulsoup4==4.14.3
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
cryptography==46.0.7
curl_cffi==0.15.0
distro==1.9.0
docstring_parser==0.17.0
frozendict==2.4.7
frozenlist==1.8.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.14.0
llvmlite==0.44.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
msgpack==1.1.2
multidict==6.7.1
multitasking==0.0.12
numba==0.61.2
numpy==2.2.6
pandas==3.0.2
pandas-ta==0.4.71b0
peewee==4.0.4
platformdirs==4.9.6
propcache==0.4.1
protobuf==7.34.1
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.20.0
PyJWT==2.12.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-http-client==3.3.7
pytz==2026.1.post1
requests==2.33.1
rich==14.3.3
sendgrid==6.12.5
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.3
sseclient-py==1.9.0
tqdm==4.67.3
twilio==9.10.4
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
websockets==16.0
Werkzeug==3.1.8
yarl==1.23.0
yfinance==1.2.1
